% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cool_matlab.R
\name{cool_matlab}
\alias{cool_matlab}
\title{Color palette for magenta-to-cyan colors (mimicing "cool" palette in matlab)}
\usage{
cool_matlab(x = 256)
}
\arguments{
\item{x}{}
}
\value{
color pallette for ggplot
}
\description{
Color palette for magenta-to-cyan colors (mimicing "cool" palette in matlab)
}
\examples{
cool_matlab()
}
\keyword{internal}
