% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_coverage.R
\name{find_coverage}
\alias{find_coverage}
\title{Find coverage of ecological community surveys at 1-k samples}
\usage{
find_coverage(occupancy, k = 1000)
}
\arguments{
\item{occupancy}{vector of average occupancy of all species in a community (>0)}

\item{k}{max number of samples to test. Coverage should asymptote at a large number of samples.}
}
\value{
vector of coverage achieved in k samples of an ecological community
}
\description{
Find coverage of ecological community surveys at 1-k samples
}
\keyword{internal}
