% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_eff_sites.R
\name{find_eff_sites}
\alias{find_eff_sites}
\title{Find effective number of sites needed to reach target coverage}
\usage{
find_eff_sites(coverage_vec, target = c(1))
}
\arguments{
\item{coverage_vec}{a vector of coverage values}

\item{target}{coverage value to achieve with n sites}
}
\value{
number of sites necessary to achieve a target coverage value
}
\description{
Find effective number of sites needed to reach target coverage
}
\keyword{internal}
