% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_target_detection_rate.R
\name{find_target_detection_rate}
\alias{find_target_detection_rate}
\title{Find target detection rate}
\usage{
find_target_detection_rate(matrix)
}
\arguments{
\item{matrix}{}
}
\value{
value of coverage at which we've found sufficient species in the community.
}
\description{
This function finds the target detection rate (coverage) that indicates that we have sampled "all" of the species in the community. Because coverage is asymptotic at 100, here, we find the point at which coverage indicates we have fewer than half of one species left to find.
}
\examples{
if(FALSE){find_target_detection_rate(pilot_single_trt)}
}
