% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_single_trt.R
\name{subsample_single_trt}
\alias{subsample_single_trt}
\title{Subsample bootstrapped communities to assess effect of sampling effort on detectable effect size}
\usage{
subsample_single_trt(
  boots,
  pilot,
  power = c(80),
  target_eff_size = NULL,
  cost_per_sample = NULL,
  seed = NULL
)
}
\arguments{
\item{pilot}{original pilot study. Identical to the pilot input in `bootstrap_single_trt`}

\item{power}{statistical power with which the user wants to detect richness change (defaults to 80)}

\item{target_eff_size}{if supplied, the level of richness change the user wants to detect (in log2 ratio units).}

\item{cost_per_sample}{cost per unit sample (linear only), which will add an axis to output plots for total cost per unit power.}

\item{seed}{random seed. Defaults to 1 so repeat runs will be identical, but since simulations rely on random draws, changing the seed will result in different answers.}

\item{data}{tibble of bootstrapped communities across multiple effect sizes. Output from `bootstrap_single_trt`.}
}
\value{
list of plots.
}
\description{
Subsample bootstrapped communities to assess effect of sampling effort on detectable effect size
}
\examples{
if(FALSE){subsample_single_trt(boots = boots, pilot = pilot_single_trt)}
}
