% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_two_trts.R
\name{subsample_two_trts}
\alias{subsample_two_trts}
\title{Subsample bootstrapped communities from two treatments}
\usage{
subsample_two_trts(
  data,
  pilot,
  power = c(80),
  target_eff_size = NULL,
  cost_per_sample = NULL,
  seed = 1
)
}
\arguments{
\item{data}{bootstrap tibble from `bootstrap_two_trt`}

\item{pilot}{species-by-site matrix from pilot survey with a column for treatment}

\item{power}{statistical power with which the user wants to detect richness change (defaults to 80)}

\item{target_eff_size}{if supplied, the level of richness change the user wants to detect (in log2 ratio units).}

\item{cost_per_sample}{cost per unit sample (linear only), which will add an axis to output plots for total cost per unit power.}

\item{seed}{random seed. Defaults to 1 so repeat runs will be identical, but since simulations rely on random draws, changing the seed will result in different answers.}
}
\value{
list of plots.
}
\description{
Subsample bootstrapped communities from two treatments
}
\examples{
if(FALSE) subsample_two_trts(boots2)
}
